+++ ropetest/refactor/patchedasttest.py
@@ -37,6 +37,60 @@ class PatchedASTTest(unittest.TestCase):
             "Expr",
         ])
 
+    def assert_function_def_has_one_type_var(self, checker):
+        checker.check_children(
+            "FunctionDef",
+            [
+                "def",
+                " ",
+                "foo",
+                "",
+                "[", "", "TypeVar", "", "]",
+                "",
+                "(", "", "arguments", "", ")",
+                "",
+                ":",
+                "\n    ",
+                "Pass",
+            ],
+        )
+
+    def assert_function_def_has_one_type_var_tuple(self, checker):
+        checker.check_children(
+            "FunctionDef",
+            [
+                "def",
+                " ",
+                "foo",
+                "",
+                "[", "", "TypeVarTuple", "", "]",
+                "",
+                "(", "", "arguments", "", ")",
+                "",
+                ":",
+                "\n    ",
+                "Pass",
+            ],
+        )
+
+    def assert_function_def_has_one_param_spec(self, checker):
+        checker.check_children(
+            "FunctionDef",
+            [
+                "def",
+                " ",
+                "foo",
+                "",
+                "[", "", "ParamSpec", "", "]",
+                "",
+                "(", "", "arguments", "", ")",
+                "",
+                ":",
+                "\n    ",
+                "Pass",
+            ],
+        )
+
     def test_operator_support_completeness(self):
         ast_ops = {
             n.__name__
@@ -1508,6 +1562,192 @@ class PatchedASTTest(unittest.TestCase):
             "MatchAs",
             "",
             "}",
+        ])
+
+    @testutils.only_for_versions_higher("3.12")
+    def test_type_alias(self):
+        source = dedent("""\
+            type Point = tuple[float, float]
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+        checker.check_children("TypeAlias", [
+            "type",
+            " ",
+            "Name",
+            " = ",
+            "Subscript",
+        ])
+
+    @testutils.only_for_versions_higher("3.12")
+    def test_type_var_simple(self):
+        source = dedent("""\
+            def foo[S, T](x):
+                pass
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+
+        checker.check_children(
+            "FunctionDef",
+            [
+                "def",
+                " ",
+                "foo",
+                "",
+                "[",
+                "",
+                "TypeVar",
+                "",
+                ",",
+                " ",
+                "TypeVar",
+                "",
+                "]",
+                "",
+                "(", "", "arguments", "", ")",
+                "",
+                ":",
+                "\n    ",
+                "Pass",
+            ],
+        )
+
+    @testutils.only_for_versions_higher("3.13")
+    def test_type_var_with_default_value(self):
+        source = dedent("""\
+            def foo[T = D](x):
+                pass
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+
+        self.assert_function_def_has_one_type_var(checker)
+
+        checker.check_children("TypeVar", [
+            "T",
+            " ",
+            "=",
+            " ",
+            "Name",
+        ])
+
+    @testutils.only_for_versions_higher("3.12")
+    def test_type_var_with_constraint(self):
+        source = dedent("""\
+            def foo[T: (A, B)](x):
+                pass
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+
+        self.assert_function_def_has_one_type_var(checker)
+
+        checker.check_children("TypeVar", [
+            "T",
+            "",
+            ":",
+            " ",
+            "Tuple",
+        ])
+
+    @testutils.only_for_versions_higher("3.13")
+    def test_type_var_with_constraint_and_default_value(self):
+        source = dedent("""\
+            def foo[T: (A, B) = D](x):
+                pass
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+
+        self.assert_function_def_has_one_type_var(checker)
+
+        checker.check_children("TypeVar", [
+            "T",
+            "",
+            ":",
+            " ",
+            "Tuple",
+            " ",
+            "=",
+            " ",
+            "Name",
+        ])
+
+    @testutils.only_for_versions_higher("3.12")
+    def test_type_var_tuple_simple(self):
+        source = dedent("""\
+            def foo[*T](x):
+                pass
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+
+        self.assert_function_def_has_one_type_var_tuple(checker)
+
+        checker.check_children("TypeVarTuple", [
+            "*",
+            "",
+            "T",
+        ])
+
+    @testutils.only_for_versions_higher("3.13")
+    def test_type_var_tuple_with_default_value(self):
+        source = dedent("""\
+            def foo[*T = (A, B)](x):
+                pass
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+
+        self.assert_function_def_has_one_type_var_tuple(checker)
+
+        checker.check_children("TypeVarTuple", [
+            "*",
+            "",
+            "T",
+            " ",
+            "=",
+            " ",
+            "Tuple",
+        ])
+
+    @testutils.only_for_versions_higher("3.12")
+    def test_param_spec_simple(self):
+        source = dedent("""\
+            def foo[**T](x):
+                pass
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+
+        self.assert_function_def_has_one_param_spec(checker)
+
+        checker.check_children("ParamSpec", [
+            "**",
+            "",
+            "T",
+        ])
+
+    @testutils.only_for_versions_higher("3.13")
+    def test_param_spec_with_default_value(self):
+        source = dedent("""\
+            def foo[**T = (A, B)](x):
+                pass
+        """)
+        ast_frag = patchedast.get_patched_ast(source, True)
+        checker = _ResultChecker(self, ast_frag)
+
+        self.assert_function_def_has_one_param_spec(checker)
+
+        checker.check_children("ParamSpec", [
+            "**",
+            "",
+            "T",
+            " ",
+            "=",
+            " ",
+            "Tuple",
         ])